Ajoute le menue en fxml

+++ src/Vue/DessinPerssonage.java
@@ -41,7 +41,7 @@ public abstract class DessinPerssonage extends Circle {
         this.posX += x;
         this.posY += y;
         Timeline timeline = new Timeline();
-        KeyFrame bouge = new KeyFrame(new Duration((this instanceof DessinIntrus) ? (float) Main.tempo / 5 : (float) Main.tempo / 2),
+        KeyFrame bouge = new KeyFrame(new Duration((this instanceof DessinIntrus) ? Main.tempo / 5 : Main.tempo / 2),
                 new KeyValue(this.centerXProperty(), posX * Main.tailleCase + Main.tailleCase + Main.tailleCase / 2),
                 new KeyValue(this.centerYProperty(), posY * Main.tailleCase + Main.tailleCase + Main.tailleCase / 2));
         timeline.getKeyFrames().add(bouge);
